requirements.txt
psutil~=7.2.2
opencv-python~=4.11.0.86
numpy~=1.26.4
tensorflow~=2.15.1
gin-config~=0.5.0
gym~=0.23.0
pygame~=2.1.3
matplotlib~=3.10.8
gym-notices~=0.1.0
defusedxml~=0.7.1
ipython~=8.38.0
pillow~=12.1.0
jupyter~=1.1.1
pip~=26.0
typing_extensions~=4.15.0
attrs~=25.4.0
protobuf>=5.28.0,<6
cryptography~=46.0.4
rsa~=4.9.1
pyasn1~=0.6.2
pyzmq~=27.1.0
pexpect~=4.9.0
cffi~=2.0.0
tornado~=6.5.4
absl-py~=2.4.0
setuptools~=80.10.2
comm~=0.2.3
gast~=0.7.0
grpcio~=1.76.0
h5py~=3.15.1
parso~=0.8.5
jedi~=0.19.2
platformdirs~=4.5.1
babel~=2.18.0
flexparser~=0.4
flexcache~=0.3
Pint~=0.24.4
packaging~=26.0
pytz~=2024.2
wrapt~=1.14.2
libclang~=18.1.1
keras~=2.15.0
six~=1.17.0
scipy~=1.17.0
pandas~=2.3.3
requests~=2.32.5
portpicker~=1.6.0
overrides~=7.7.0
charset-normalizer~=3.4.4
google-pasta~=0.2.0
threadpoolctl~=3.6.0
pyasn1_modules~=0.4.2
certifi~=2026.1.4
urllib3~=2.6.3
joblib~=1.5.3
cloudpickle~=3.1.2
MarkupSafe~=3.0.3
python-dateutil~=2.9.0.post0
termcolor~=3.3.0
debugpy~=1.8.20
traitlets~=5.14.3
decorator~=5.2.1
Pygments~=2.19.2
stack-data~=0.6.3
matplotlib-inline~=0.2.1
jupyter_client~=8.8.0
prompt_toolkit~=3.0.52
wcwidth~=0.5.3
ptyprocess~=0.7.0
seaborn~=0.13.2
cycler~=0.12.1
ipywidgets~=8.1.8
scikit-learn~=1.8.0
holidays~=0.56
Markdown~=3.10.1
PyYAML~=6.0.3
oauthlib~=3.3.1
idna~=3.11
asttokens~=3.0.1
contourpy~=1.3.3
executing~=2.2.1
fonttools~=4.61.1
ipykernel~=6.31.0
jupyter_core~=5.9.1
ml-dtypes~=0.3.2
pure_eval~=0.2.3
pyparsing~=3.3.2
Jinja2~=3.1.6
kiwisolver~=1.4.9
opt_einsum~=3.4.0
wheel~=0.46.3
tensorboard~=2.15.2
flatbuffers~=25.12.19
astunparse~=1.6.3
Werkzeug~=3.1.5
google-auth-oauthlib~=1.2.4
requests-oauthlib~=2.0.0
smart_control~=0.1.0
mediapy~=1.2.5
python-dotenv~=1.2.1
bidict~=0.23.1